+++ SU2_CFD/include/numerics_structure.inl
@@ -576,13 +576,3 @@ inline void CNumerics::ComputeResidual(su2double **val_Jacobian_i, su2double *va
                   su2double **val_Jacobian_j, su2double *val_Jacobian_muj, su2double ***val_Jacobian_gradj, CConfig *config) { }
 
 inline void CNumerics::SetTauWall(su2double val_tauwall_i, su2double val_tauwall_j) { }
-
-inline void CAvgGrad_Flow::SetTauWall(su2double val_tauwall_i, su2double val_tauwall_j) {
-  TauWall_i = val_tauwall_i;
-  TauWall_j = val_tauwall_j;
-}
-
-inline void CAvgGradCorrected_Flow::SetTauWall(su2double val_tauwall_i, su2double val_tauwall_j) {
-  TauWall_i = val_tauwall_i;
-  TauWall_j = val_tauwall_j;
-}
