+++ SU2_CFD/src/driver_structure.cpp
@@ -36,6 +36,7 @@
  */
 
 #include "../include/driver_structure.hpp"
+#include "../include/numerics_direct_mean.hpp"
 #include "../include/definition_structure.hpp"
 
 #ifdef VTUNEPROF
